moved encode_uid and decode_uid to users/auth/utils and send_password_reset to users/auth/services
