+++ {{cookiecutter.github_repository}}/tests/integration/test_auth.py
@@ -6,7 +6,7 @@ import pytest
 from django.urls import reverse
 from django.contrib.auth.tokens import default_token_generator
 
-from {{cookiecutter.main_module}}.users.services import encode_uid
+from {{cookiecutter.main_module}}.users.auth.utils import encode_uid
 from .. import factories as f
 
 pytestmark = pytest.mark.django_db
