+++ {{cookiecutter.github_repository}}/{{cookiecutter.main_module}}/users/auth/utils.py
@@ -1,0 +1,11 @@
+# Third Party Stuff
+from django.utils.encoding import force_bytes, force_text
+from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
+
+
+def encode_uid(pk):
+    return urlsafe_base64_encode(force_bytes(pk)).decode()
+
+
+def decode_uid(pk):
+    return force_text(urlsafe_base64_decode(pk))
